third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/form-action-resubmission-main-page-expected.txt

CONSOLE ERROR: Refused to send form data to 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/form-action-resubmission-main-page-callee.html' because it violates the following Content Security Policy directive: "form-action 'none'".

This is a testharness.js-based test.
PASS A form resubmission in the main frame should be blocked
Harness: the test ran to completion.
